requirements.txt
Django==1.7.7
dj-database-url==0.3.0
dj-static==0.0.6
gunicorn==19.3.0
pystache==0.5.4
static==1.1.1
static3==0.5.1
wsgiref==0.1.2
psycopg2==2.6
django-extensions==1.5.2
django-storages==1.1.8
boto==2.36.0
ipython==3.0.0
django-mail-queue==2.2.0
django-compressor==1.4
django-bootstrap-form==3.2
django-sekizai==0.8.1
